Added debug/print modes
Change Debug =1 to debug code, change print =1 to display normal print.
Added a print roll function and changed it to print horizontally as it should. Started work on reels, have a way to make it work need to figure out payout as it would be different. Added a total bonus so you can see the total winnings after a bonus